notifications.js / checkout popup & profile popup
When requesting to fork , it would only give me the older extension which uses json files. I still want send a pull request however so the code I've been working on is easier to access. I created a checkout popup, that only comes after the user is earning and has met the requirements. I also have a function that creates the profile popup which congratulates them , let them know of their soulsmiles due to their last purchase, and asks if they want to donate.