added in popups
I created the checkout popup to come at the checkout page instead of the earnings reminder. I also have the profile popup as well that shows at the homepage after being delayed for 30 seconds . The profile popup allows you to donate soulsmiles and it should have a the amount of soulsmiles in wallet. 

There are issues with  importing of the account.js and firebase files, but I wanted to request a pull so you can see what's wrong exactly.